requirements.txt
langchain
langchain-community
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
scikit-learn
langchain_groq

huggingface-hub==0.14.1


InstructorEmbedding==1.0.1
sentence-transformers==2.2.2